src/backend/pyproject.toml
# Copyright (c) 2022, 2023 Humanitarian OpenStreetMap Team
#
# This file is part of FMTM.
#
#     FMTM is free software: you can redistribute it and/or modify
#     it under the terms of the GNU General Public License as published by
#     the Free Software Foundation, either version 3 of the License, or
#     (at your option) any later version.
#
#     FMTM is distributed in the hope that it will be useful,
#     but WITHOUT ANY WARRANTY; without even the implied warranty of
#     MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
#     GNU General Public License for more details.
#
#     You should have received a copy of the GNU General Public License
#     along with FMTM.  If not, see <https:#www.gnu.org/licenses/>.
#

[project]
name = "fmtm"
dynamic = ["version"]
description = "Field Mapping Tasking Manager - coordinated field mapping."
authors = [
    {name = "HOTOSM", email = "sysadmin@hotosm.org"},
]
dependencies = [
    "uvicorn==0.32.0",
    "uvloop==0.21.0",
    "httptools==0.6.4",
    "fastapi==0.115.3",
    "asgi-lifespan==2.1.0",
    "pydantic==2.9.2",
    "pydantic-settings==2.6.0",
    "geojson-pydantic==1.1.2",
    "python-multipart==0.0.12",
    "psycopg[pool]>=3.2.3",
    "geojson==3.1.0",
    "shapely==2.0.6",
    "pyxform==2.1.1",
    "loguru==0.7.2",
    "minio==7.2.9",
    "pyproj==3.7.0",
    "asgiref==3.8.1",
    "sozipfile==0.3.2",
    "cryptography==43.0.3",
    "pyjwt==2.9.0",
    "osm-fieldwork==0.16.9",
    "osm-login-python==2.0.0",
    "osm-rawdata==0.4.0",
    "fmtm-splitter==1.4.0",
]
requires-python = ">=3.12"
readme = "../../README.md"
license = {text = "GPL-3.0-only"}

[project.optional-dependencies]
dev = [
    "pre-commit>=4.0.1",
    "commitizen>=3.30.0",
]
test = [
    "pytest==8.3.3",
    "pytest-asyncio==0.24.0",
    "httpx==0.27.2",
    "pyinstrument==5.0.0",
    "coverage>=7.6.4",
    "coverage-badge>=1.1.2",
]
debug = [
    "debugpy>=1.8.7",
]
docs = [
    "mkdocs==1.6.1",
    "mkdocs-material==9.5.42",
    "mkdocstrings-python==1.12.2",
    "mkdocs-exclude==1.0.2",
    "python-dotenv==1.0.1",
    "mkdocs-git-revision-date-localized-plugin>=1.3.0",
    "mdx-truly-sane-lists>=1.3",
]
monitoring = [
    "sentry-sdk[opentelemetry]>=2.17.0",
    #"opentelemetry-sdk>=1.27.0",
    #"opentelemetry-api>=1.27.0",
    #"opentelemetry-exporter-otlp>=1.27.0",
    "opentelemetry-instrumentation-fastapi>=0.48b0",
    "opentelemetry-instrumentation-logging>=0.48b0",
    "opentelemetry-instrumentation-psycopg2>=0.48b0",
    "opentelemetry-instrumentation-requests>=0.48b0",
]

[build-system]
requires = ["pdm-pep517==1.1.4"]
build-backend = "pdm.pep517.api"

[tool.pdm]
source-includes = ["../../CHANGELOG.md", "../../LICENSE", "../../README.md"]
version = {from = "app/__version__.py"}

[tool.ruff]
fix = true
line-length = 88
target-version = "py311"
exclude = [
    ".ruff_cache",
    "__pypackages__",
    "app/__version__.py",
]
[tool.ruff.lint]
select = ["I", "E", "W", "D", "B", "F", "N", "Q"]
ignore = ["N805", "B008"]
[tool.ruff.lint.pydocstyle]
convention = "google"
[tool.ruff.lint.flake8-bugbear]
extend-immutable-calls = [
    "fastapi.Depends",
    "fastapi.params.Depends",
    "fastapi.Query",
    "fastapi.params.Query",
    "fastapi.File",
    "fastapi.Security",
]

[tool.pyright]
extraPaths = ["__pypackages__/3.12/lib/"]

[tool.pytest.ini_options]
addopts = "-ra -q"
testpaths = [
    "tests",
]
asyncio_mode="auto"
asyncio_default_fixture_loop_scope="session"

[tool.commitizen]
name = "cz_conventional_commits"
version = "2024.4.1"
version_files = [
    "pyproject.toml:version",
    "app/__version__.py",
    "../frontend/package.json:version",
    "../../chart/Chart.yaml:appVersion",
]
changelog_file = "../../CHANGELOG.md"
update_changelog_on_bump = true

[tool.sqlfluff.core]
dialect = "postgres"
large_file_skip_byte_limit = 30000  # Required to process fmtm_base_schema.sql
exclude_rules = "CP05, RF02"  # Avoid capitalisation of enums

[tool.codespell]
skip = "contrib/*.py,*languages_and_countries.py,*pnpm-lock.yaml,*CHANGELOG.md,src/mapper/src/lib/migrations/*"
ignore-words-list = "ProjectIn"
write-changes = true
